Introduce a symbols API
Rework how we store and search symbols for relocation, abstracting away
most of the detail, in order to be able to use various underlying
data types. As relocation seems to take most of the runtime based on
initial observations, this is a first step in order to compare against
various backend data types, rather than using an ad-hoc implementation.